requirements.txt
alabaster==0.7.12
amqp==5.0.2
arabic-reshaper==2.1.1
asgiref==3.3.1
Babel==2.9.0
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.6
django-countries==7.0
django-crispy-forms==1.10.0
django-dajax==0.9.2
django-dajaxice==0.7
django-extensions==3.1.0
django-formtools==2.2
django-heroku==0.3.1
django-javascript-settings==1.1.2
django-storages==1.11.1
docutils==0.16
future==0.18.2
fuzzywuzzy==0.18.0
google-api-core==1.25.1
google-auth==1.25.0
google-cloud-core==1.6.0
google-cloud-storage==1.35.1
google-cloud-vision==2.1.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
imagesize==1.2.0
importlib-metadata==3.4.0
Jinja2==2.11.2
kombu==5.0.2
libcst==0.3.16
livereload==2.6.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.8
Pillow==8.1.0
prompt-toolkit==3.0.10
proto-plus==1.13.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.1
pytz==2019.3
PyYAML==5.4.1
reportlab==3.5.59
requests==2.25.1
rsa==4.7
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.4.3
sphinx-autobuild==2020.9.1
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
tornado==6.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
zipp==3.4.0
